+++ src/api/queries/usePatientSummary.ts
@@ -1,8 +1,5 @@
-import axios from 'axios'
-import { useQuery } from 'react-query'
-import { PaginatedResponse } from '../../types/paginatedResponse'
-import { createQueryKey } from '../../utils/url'
 import { FacilitySummaryResponse } from './useFacilitySummary'
+import { getGenericQueryHook } from './utils'
 
 const PATIENT_SUMMARY_KEY = 'patientSummeryKey'
 
@@ -40,19 +37,7 @@ export interface PatientSummaryQuery {
   facility?: string
 }
 
-export const usePatientSummary = (query: PatientSummaryQuery, enabled = true) =>
-  useQuery(
-    createQueryKey(PATIENT_SUMMARY_KEY, query),
-    () =>
-      axios
-        .get<PaginatedResponse<PatientSummaryResponse[]>>(
-          'https://careapi.coronasafe.in/api/v1/patient_summary/',
-          {
-            params: query,
-          }
-        )
-        .then((d) => d.data),
-    {
-      enabled,
-    }
-  )
+export default getGenericQueryHook<PatientSummaryQuery, PatientSummaryResponse>(
+  PATIENT_SUMMARY_KEY,
+  '/api/v1/patient_summary/'
+)
