+++ src/api/queries/useTriageSummary.ts
@@ -1,7 +1,3 @@
-import axios from 'axios'
-import { useQuery } from 'react-query'
-import { PaginatedResponse } from '../../types/paginatedResponse'
-import { createQueryKey } from '../../utils/url'
 import { FacilitySummaryResponse } from './useFacilitySummary'
 import { getGenericQueryHook } from './utils'
 
@@ -39,5 +35,5 @@ export interface TriageSummaryData {
 
 export default getGenericQueryHook<TriageSummaryQuery, TriageSummaryResponse>(
   TRIAGE_SUMMARY_KEY,
-  'api/v1/triage_summary'
+  'api/v1/triage_summary/'
 )
